// learning_system_frontend/src/pages/inventory/add_to_inventory/types.ts
export interface Upload_compo_details_int {
  componentTitle: string;
  availableQuantity: string;
  allocatedQuantity: string;
  isAllowedToSelect: boolean;
  isCostly: boolean;
  refrenceLink1: string;
  refrenceLink2: string;
  approxAmount: string;
  componentImageUrl: string;
  componentDescription: string;
}